Fixed memory allocation error.
Images for sprites were being killed before the sprites leading to internal errors. To solve this I moved the KillResource code out into the gui manager. However, this then lead to errors in the action bar as it wasn't destroying the resources used for it's own buttons.

To fix this I also implemented code within the action bars destructor to destroy the resource for each button it creates.

Added an example button that remains in the exact center of the screen as another demonstration of GUI coords.